made a few changes
taking a different approach, first im gin to build the whole project for a single type of command, doing that will get me a skeleton readt with a fully functional command. This would let me write the complete infrastructure for the DataBase and hence would help in writing and scaling harixontally of other functions.